Replace files, make it compile
